AUT-4248: Move CommonTestVariables from frontend-api to shared
These variables have uses across projects here and in particular the MFAMethods set up here will be used in our shared tests, not just frontend, in upcoming commits
