+++ frontend-api/src/test/java/uk/gov/di/authentication/frontendapi/helpers/ApiGatewayProxyRequestHelper.java
@@ -5,7 +5,7 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent
 
 import java.util.Map;
 
-import static uk.gov.di.authentication.frontendapi.helpers.CommonTestVariables.IP_ADDRESS;
+import static uk.gov.di.authentication.shared.helpers.CommonTestVariables.IP_ADDRESS;
 import static uk.gov.di.authentication.sharedtest.helper.RequestEventHelper.contextWithSourceIp;
 
 public class ApiGatewayProxyRequestHelper {
